Codes on function pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -9,7 +9,8 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	unsigned long int i;
+
 	if (array == NULL)
 		return;
 	if (size <= 0)
